Update CNAME
These two entries tell GitHub to redirect any requests to jacoblite.github.io to the domain name. There are two entries to ensure that jacoblite.com will also redirect to www.jacoblite.com (if a site with no www prefix is preferred, you can switch the order of the domains).